backend/requirements.txt
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
debugpy==1.8.12
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.2.0
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.4
MouseInfo==0.1.3
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.2.4
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pika==1.3.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
PyMySQL==1.1.2
pyodbc==5.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.27.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.14.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.23.0
